src/main/webapp/app/entities/shipment-cart/delete/shipment-cart-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ShipmentCartService } from '../service/shipment-cart.service';

import { ShipmentCartDeleteDialogComponent } from './shipment-cart-delete-dialog.component';

describe('ShipmentCart Management Delete Component', () => {
  let comp: ShipmentCartDeleteDialogComponent;
  let fixture: ComponentFixture<ShipmentCartDeleteDialogComponent>;
  let service: ShipmentCartService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [ShipmentCartDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ShipmentCartDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ShipmentCartDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ShipmentCartService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
